requirements.txt
# Core dependencies for Layered Context Graph Master Processor
# Natural language processing
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0

# Graph processing
networkx>=2.8.0
scipy>=1.9.0
numpy>=1.21.0

# Ollama integration (optional)
ollama>=0.1.0

# Utilities
pathlib2>=2.3.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and testing
pytest>=7.0.0
jupyter>=1.0.0

gguf
llama-cpp-python
ipykernel
h5py
